All updates

+++ Microservices/Documents.API/Documents.API/Documents.API.Application/Command/Update/UpdateCommandHandler.cs
@@ -14,41 +14,51 @@ namespace Documents.API.Application.Command.Update
         private readonly IBlobServices _blobService;
         private readonly BlobStorageSettings _blobStorageSettings;
         private readonly IPhotosRepository _documentsRepository;
+        private readonly IPhotoService _photoService;
 
         public UpdateCommandHandler(IBlobServices blobService, IOptions<BlobStorageSettings> blobStorageSettings,
-            IPhotosRepository documentsRepository)
+            IPhotosRepository documentsRepository,IPhotoService photoService)
         {
             _blobService = blobService;
             _blobStorageSettings = blobStorageSettings.Value;
             _documentsRepository = documentsRepository;
+            _photoService =photoService;
         }
         public async Task<long> Handle(UpdateCommand request, CancellationToken cancellationToken)
         {
             var document = await _documentsRepository.GetAsync(request.FileDto.PhotoId, cancellationToken);
-            // _documentsRepository.DeleteAsync(document.Id, cancellationToken);
-            await _blobService.DeleteAsync(document.ContainerName, document.Path, cancellationToken);
-            string directory = System.IO.Path.GetDirectoryName(document.Path);
-            string newPath = System.IO.Path.Combine(directory, request.FileDto.File.FileName);
-            document.Path = newPath;
-            document.FileName = request.FileDto.File.FileName;
-           
-            var pathTemplate = GetTemplatePath(request.FileDto.Subject);
-           
-            try
+            if (document != null)
             {
-                await using (var filestream = request.FileDto.File.OpenReadStream())
+
+                // _documentsRepository.DeleteAsync(document.Id, cancellationToken);
+                await _blobService.DeleteAsync(document.ContainerName, document.Path, cancellationToken);
+                string directory = System.IO.Path.GetDirectoryName(document.Path);
+                string newPath = System.IO.Path.Combine(directory, request.FileDto.File.FileName);
+                document.Path = newPath;
+                document.FileName = request.FileDto.File.FileName;
+
+                var pathTemplate = GetTemplatePath(request.FileDto.Subject);
+
+                try
                 {
-                    await _blobService.UploadAsync(document.ContainerName, newPath, filestream, cancellationToken);
+                    await using (var filestream = request.FileDto.File.OpenReadStream())
+                    {
+                        await _blobService.UploadAsync(document.ContainerName, newPath, filestream, cancellationToken);
+                    }
+                    
+                }
+                catch (Exception ex)
+                {
+                    await _blobService.DeleteAsync(document.ContainerName, newPath, cancellationToken);
                 }
 
+                await _documentsRepository.UpdateAsync(document, cancellationToken);
+                return document.Id;
             }
-            catch (Exception ex)
-            {
-                await _blobService.DeleteAsync(document.ContainerName, newPath, cancellationToken);
+            else
+            { 
+                throw new Exception("There are not any photo for this entity");
             }
-
-            await  _documentsRepository.UpdateAsync(document, cancellationToken);
-            return document.Id;
         }
         private string GetTemplatePath(SubjectUpdate subject)
         {
